+++ src/entities/Dish/index.ts
@@ -1,3 +1,4 @@
 export { DishCard } from './ui/DishCard';
 export { getTagColors } from './lib/helpers/getTagColors';
 export { pickDishData } from './lib/helpers/pickDish';
+export type { IDishData } from './model/types';
