+++ src/pages/Home/ui/home-page/index.tsx
@@ -4,7 +4,7 @@ import { HeroSection } from '../hero-section';
 import { Stats } from '../stats';
 import { AboutMobile } from '../about-mobile';
 import { TopRestaurant } from '../top-restaurants';
-import { TopDishes } from '../top-dishes';
+import { TopDishes } from '@widgets/top-dishes';
 
 export const Home = () => {
     return (
