+++ src/shared/const/colors/tagsColors.ts
