+++ src/widgets/top-dishes/index.ts
@@ -1,0 +1,1 @@
+export { TopDishes } from './ui/index';
