+++ src/widgets/top-dishes/ui/styles.module.scss
